add photo step to context
